# FIX: Contour lines rendered incorrectly when closed loops