--- a/lib/matplotlib/contour.py
+++ b/lib/matplotlib/contour.py
@@ -1037,6 +1037,16 @@ class ContourSet(cm.ScalarMappable, ContourLabeler):
         return (lowers, uppers)
 
     def _make_paths(self, segs, kinds):
+        """
+        Create and return Path objects for the specified segments and optional
+        kind codes.  segs is a list of numpy arrays, each array is either a
+        closed line loop or open line strip of 2D points with a shape of
+        (npoints, 2).  kinds is either None or a list (with the same length as
+        segs) of numpy arrays, each array is of shape (npoints,) and contains
+        the kinds codes for the corresponding line in segs.  If kinds is None
+        then the Path constructor creates the kind codes assuming that the line
+        is an open strip.
+        """
         if kinds is None:
             return [mpath.Path(seg) for seg in segs]
         else: