--- a/lib/matplotlib/tests/test_contour.py
+++ b/lib/matplotlib/tests/test_contour.py
@@ -398,3 +398,15 @@ def test_contour_linewidth(
 def test_label_nonagg():
     # This should not crash even if the canvas doesn't have a get_renderer().
     plt.clabel(plt.contour([[1, 2], [3, 4]]))
+
+
+@image_comparison(baseline_images=['contour_closed_line_loop'],
+                  extensions=['png'], remove_text=True)
+def test_contour_closed_line_loop():
+    # github issue 19568.
+    z = [[0, 0, 0], [0, 2, 0], [0, 0, 0], [2, 1, 2]]
+
+    fig, ax = plt.subplots(figsize=(2, 2))
+    ax.contour(z, [0.5], linewidths=[20], alpha=0.7)
+    ax.set_xlim(-0.1, 2.1)
+    ax.set_ylim(-0.1, 3.1)