--- a/lib/matplotlib/tri/tricontour.py
+++ b/lib/matplotlib/tri/tricontour.py
@@ -33,7 +33,7 @@ class TriContourSet(ContourSet):
         Process args and kwargs.
         """
         if isinstance(args[0], TriContourSet):
-            C = args[0].cppContourGenerator
+            C = args[0]._contour_generator
             if self.levels is None:
                 self.levels = args[0].levels
         else:
@@ -43,28 +43,8 @@ class TriContourSet(ContourSet):
             self._mins = [tri.x.min(), tri.y.min()]
             self._maxs = [tri.x.max(), tri.y.max()]
 
-        self.cppContourGenerator = C
+        self._contour_generator = C
         return kwargs
-
-    def _get_allsegs_and_allkinds(self):
-        """
-        Create and return allsegs and allkinds by calling underlying C code.
-        """
-        allsegs = []
-        if self.filled:
-            lowers, uppers = self._get_lowers_and_uppers()
-            allkinds = []
-            for lower, upper in zip(lowers, uppers):
-                segs, kinds = self.cppContourGenerator.create_filled_contour(
-                    lower, upper)
-                allsegs.append([segs])
-                allkinds.append([kinds])
-        else:
-            allkinds = None
-            for level in self.levels:
-                segs = self.cppContourGenerator.create_contour(level)
-                allsegs.append(segs)
-        return allsegs, allkinds
 
     def _contour_args(self, args, kwargs):
         if self.filled: