… lib/matplotlib/tests/baseline_images/test_axes/contour_colorbar.pdf
… lib/matplotlib/tests/baseline_images/test_axes/contour_colorbar.png
… lib/matplotlib/tests/baseline_images/test_axes/contour_colorbar.svg
… lib/matplotlib/tests/baseline_images/test_axes/imshow_clip.pdf
… lib/matplotlib/tests/baseline_images/test_axes/imshow_clip.png
… lib/matplotlib/tests/baseline_images/test_axes/imshow_clip.svg
… lib/matplotlib/tests/baseline_images/test_contour/contour_addlines.png
… lib/matplotlib/tests/baseline_images/test_contour/contour_closed_line_loop.png
… lib/matplotlib/tests/baseline_images/test_contour/contour_line_start_on_corner_edge.png
… lib/matplotlib/tests/baseline_images/test_contour/contour_manual.png
… lib/matplotlib/tests/baseline_images/test_lines/line_collection_dashes.pdf
… lib/matplotlib/tests/baseline_images/test_lines/line_collection_dashes.svg
… lib/matplotlib/tests/baseline_images/test_patheffects/tickedstroke.png
… lib/matplotlib/tests/baseline_images/test_triangulation/tri_smooth_contouring.png
… lib/matplotlib/tests/test_triangulation.py
… lib/mpl_toolkits/tests/baseline_images/test_axisartist_axislines/ParasiteAxesAuxTrans_meshplot.png
… src/tri/_tri.cpp