## PR Summary

Fixes issue #19568 so that closed contour line loops are rendered correctly.

Solution is to use `PathCollection`s rather than `LineCollection`s, the same as for filled contours.  Added a specific test for this, and updated the 11 test images affected by the change.

To test the performance impact I've used the following code:
```python
import matplotlib
matplotlib.use('Agg')
import matplotlib.pyplot as plt
import numpy as np

shape = (500, 1000)
x, y = np.meshgrid(np.linspace(0.0, 1.0, shape[1]), np.linspace(0.0, 1.0, shape[0]))

np.random.seed(9876)
z = np.random.uniform(size=shape)
z = np.ma.array(z, mask=np.random.uniform(size=shape) < 0.1)

fig = plt.figure(figsize=(80, 40))
ax = fig.add_axes([0.0, 0.0, 1.0, 1.0])
ax.contour(x, y, z)
fig.savefig('contour_lines.png')
```
which creates an 8000x4000 PNG with 117,000 line loops and 201,000 line strips.  I timed it before and after this PR, 10 times each (yes I should have used `timeit` but I didn't) and the walltimes are:

- Before PR: minimum = 11.994 s, mean = 12.081 s, stddev = 0.046 s
- After PR: minimum = 12.273 s, mean = 12.301 s, stddev = 0.041 s

So minimum run time is 2.3% slower, mean run time is 1.8% slower.  Let's call it 2% slower (on my new but cheap laptop).  It is slower for 2 reasons:
1.  It is rendering more pixels.
2. Time taken to determine if a line segment is closed or not, which here is performed ~318k times.

Item 2 could be made faster by moving the calculations from Python to C++.  I would prefer to not to that now and instead consider it as part of a larger study into improving contouring performance.

## PR Checklist

- [x] Has pytest style unit tests (and `pytest` passes).
- [x] Is [Flake 8](https://flake8.pycqa.org/en/latest/) compliant (run `flake8` on changed files to check).
- [ ] New features are documented, with examples if plot related.
- [ ] Documentation is sphinx and numpydoc compliant (the docs should [build](https://matplotlib.org/devel/documenting_mpl.html#building-the-docs) without error).
- [ ] Conforms to Matplotlib style conventions (install `flake8-docstrings` and run `flake8 --docstring-convention=all`).
- [ ] New features have an entry in `doc/users/next_whats_new/` (follow instructions in README.rst there).
- [ ] API changes documented in `doc/api/next_api_changes/` (follow instructions in README.rst there).