@QuLogic 2021-03-03
As you mentioned in the issue, the C++ code creates the codes, and it seems like those are returned in both cases:
https://github.com/matplotlib/matplotlib/blob/0b9a0fe25f4dde83647e5259c99d48581b01bfc9/lib/matplotlib/contour.py#L826
Can you not use that in the `_make_paths` call?