@QuLogic 2021-03-03
Ah no, I see it doesn't:
https://github.com/matplotlib/matplotlib/blob/0b9a0fe25f4dde83647e5259c99d48581b01bfc9/lib/matplotlib/contour.py#L1399-L1416
But wouldn't it be less trouble to go to that directly?