@ianthomas23 2021-03-03
>  But wouldn't it be less trouble to go to that directly?

Not really.  Returning the kind codes from C++ would save us a few lines of Python but make the C++ more complicated, and it needs to be in both `_contour.cpp` and `_tri.cpp`.  Those C++ changes will be obsolete in a few months when I address the performance issues in a coherent way.

[Aside: The major performance issue for contour lines in the example here is that Python/C++ create 318k numpy arrays of vertices and 318k numpy arrays of kind codes which are passed to the renderer for consumption.  That is a lot of memory thrashing.  It is only done this way because it was necessary for the 'new' (2014) contour algorithm to match the 'old' (2005) one.  After 7 years I think we can be brave enough to change this to, for example, a smaller number of larger numpy arrays.  But that needs to be done carefully and is outside the scope of this PR.]

As you have asked, I have added a commit that returns the kind codes from the C++.  Interestingly it is marginally slower than the Python-only solution but my timings are not comprehensive.  I'd be happy to roll back the PR to the Python-only changes if that is preferred.  It is certainly simpler and lower risk.