@ianthomas23 2021-03-10
Tests are passing now that I have modified the examples and mplot3d in line with the other changes.  I've also added some new tests to increase the coverage.

There are two possible solutions here:

1. Simple Python-only fix (commit 1b15a11),
2. More complicated but more correct C++ fix (latest commit).

Either would be acceptable in terms of fixing the problem. I slightly favour of the C++ approach but could be persuaded either way. It is riskier than the Python-only solution in that there will almost certainly be some glitches caused by it, but I have the time and motivation to deal with these so now seems as good a time as any.