@jklymak 2021-03-22
I guess I don't understand the tradeoffs in putting the work in the C++ code.  Certainly there are a lot of changes, but they seem to be mostly just adding codes to the structure?  The python changes are simpler, so what is the goal of pushing it down to C++?