@ianthomas23 2021-03-22
The two implementations (Python and C++) are functionally equivalent and produce identical output.  The difference is where the codes are calculated.

The benefit of the C++ implementation is that it is faster at determining if each contour line is a closed line loop or open line strip.  However, this performance benefit is not a practical benefit as the time taken to create lots of numpy arrays swamps everything else.  When I eventually improve the performance of contouring this will become important, but it isn't yet.