@ianthomas23 2021-03-23
The git diff isn't hugely helpful here; looking at the changed C++ is probably better as the relationship between the vertices and codes arrays is more obvious.  But yes, I am happy to add more comments.