@ianthomas23 2021-04-07
@jklymak I've added more comments to `_contour.cpp` and `_tri.cpp` to explain the data returned from C++ to python, and also in the python function `_make_paths`.